refactor(alerting): consolidate into alerting_manage_alertmanager
