+++ README.md
@@ -369,9 +369,9 @@ Scopes define the specific resources that permissions apply to. Each action requ
 | `list_snowflake_tables`           | Snowflake*                | List tables in a Snowflake database/schema via INFORMATION_SCHEMA                                            | `datasources:query`                                    | `datasources:uid:*`                                 |
 | `describe_snowflake_table`        | Snowflake*                | Get table schema (column types, nullability, defaults, comments)                                             | `datasources:query`                                    | `datasources:uid:*`                                 |
 | `query_snowflake`                 | Snowflake*                | Execute SQL queries with macro/variable substitution                                                         | `datasources:query`                                    | `datasources:uid:*`                                 |
+| `alerting_manage_alertmanager`    | Alerting                  | Manage alert groups and silences (list_alert_groups, list, get, create, update, delete)                      | `alert.instances:read` + `alert.instances:write` for mutations | Global scope                                        |
 | `alerting_manage_rules`           | Alerting                  | Manage alert rules (list, get, versions, create, update, delete)                                             | `alert.rules:read` + `alert.rules:write` for mutations | `folders:*` or `folders:uid:alerts-folder`          |
 | `alerting_manage_routing`         | Alerting                  | Manage notification policies, contact points, and time intervals                                             | `alert.notifications:read`                             | Global scope                                        |
-| `alerting_manage_silences`        | Alerting                  | Manage alerting silences (list, get, create, update, expire)                                                 | `alert.instances:read` + `alert.instances:write` for mutations | Global scope                                        |
 | `list_oncall_schedules`           | OnCall                    | List schedules from Grafana OnCall                                                                           | `grafana-oncall-app.schedules:read`                    | Plugin-specific scopes                              |
 | `get_oncall_shift`                | OnCall                    | Get details for a specific OnCall shift                                                                      | `grafana-oncall-app.schedules:read`                    | Plugin-specific scopes                              |
 | `get_current_oncall_users`        | OnCall                    | Get users currently on-call for a specific schedule                                                          | `grafana-oncall-app.schedules:read`                    | Plugin-specific scopes                              |
@@ -516,8 +516,8 @@ When `--disable-write` is enabled, the following write operations are disabled:
 - `update_incident`
 
 **Alerting Tools:**
+- `alerting_manage_alertmanager` (create, update, delete operations)
 - `alerting_manage_rules` (create, update, delete operations)
-- `alerting_manage_silences` (create, update, delete operations)
 
 **OnCall Tools:**
 - `update_alert_group`
